refactoring some code
