+++ app/src/main/java/com/example/termprojectuser/Adapter/RewardAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.termprojectuser
+package com.example.termprojectuser.Adapter
 
 import android.content.Context
 import android.view.LayoutInflater
@@ -7,9 +7,10 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.constraintlayout.widget.ConstraintLayout
-import androidx.coordinatorlayout.widget.CoordinatorLayout
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
+import com.example.termprojectuser.Entity.Menu
+import com.example.termprojectuser.R
 import com.firebase.ui.database.FirebaseRecyclerAdapter
 import com.firebase.ui.database.FirebaseRecyclerOptions
