+++ app/src/main/java/com/example/termprojectuser/Entity/Menu.kt
@@ -1,4 +1,4 @@
-package com.example.termprojectuser
+package com.example.termprojectuser.Entity
 
 import android.os.Parcel
 import android.os.Parcelable
@@ -13,19 +13,7 @@ data class Menu(
         val price: Int = -1,
         var remain: Int = -1
         ): Parcelable{
-//        companion object{
-//                fun createMenu():ArrayList<Menu>{
-//                        var menu = ArrayList<Menu>()
-//                        menu.add(Menu(R.drawable.yuzu_refresher, "Yuzu refresher", 35,10))
-//                        menu.add(Menu(R.drawable.classic_brown_sugar_milk_tea,"Classic brown sugar milk tea", 75,10 ))
-//                        menu.add(Menu(R.drawable.matcha_brown_sugar_latte, "Matcha brown sugar latte", 105,10))
-//                        menu.add(Menu(R.drawable.traditional_thai_milk_tea, "Traditional Thai milk tea", 15,10))
-//                        menu.add(Menu(R.drawable.hojicha_latte, "Hojicha latte", 40,10))
-//                        menu.add(Menu(R.drawable.caramel_macchiato, "Caramel macchiato", 40,10))
-//                        return menu
-//
-//                }
-//        }
+
         fun subtractRemain(){
                 remain--
         }
