+++ app/src/main/java/com/example/termprojectuser/Entity/Order.kt
@@ -1,4 +1,4 @@
-package com.example.termprojectuser
+package com.example.termprojectuser.Entity
 
 
 import android.os.Parcelable
