+++ app/src/main/java/com/example/termprojectuser/Entity/Queue.kt
@@ -1,0 +1,6 @@
+package com.example.termprojectuser.Entity
+
+data class Queue(
+        val orderList: Map<String, OrderList> = mapOf(),
+        val queueId: String = "Unknown"
+)
