+++ app/src/main/java/com/example/termprojectuser/Entity/RecyclerItem.kt
@@ -1,8 +1,8 @@
-package com.example.termprojectuser
+package com.example.termprojectuser.Entity
 
 sealed class RecyclerItem {
     data class Header(val typeName: String): RecyclerItem()
-    data class Product(val order:Order): RecyclerItem()
+    data class Product(val order: Order): RecyclerItem()
     companion object{
         fun transformList(orderList: ArrayList<Order>):ArrayList<RecyclerItem>{
             val groupList = orderList.groupBy {
@@ -12,14 +12,14 @@ sealed class RecyclerItem {
             val myRedeemList = ArrayList<RecyclerItem>()
             for (i in groupList.keys){
                 if (i){
-                    myRedeemList.add(RecyclerItem.Header("Reward Order"))
+                    myRedeemList.add(Header("Reward Order"))
                     for (v in groupList.getValue(i)){
-                        myRedeemList.add(RecyclerItem.Product(v))
+                        myRedeemList.add(Product(v))
                     }
                 }else{
-                    myOrderList.add(RecyclerItem.Header("Order"))
+                    myOrderList.add(Header("Order"))
                     for (v in groupList.getValue(i)){
-                        myOrderList.add(RecyclerItem.Product(v))
+                        myOrderList.add(Product(v))
                     }
                 }
             }
