+++ app/src/main/java/com/example/termprojectuser/Entity/RewardMenu.kt
@@ -1,0 +1,8 @@
+package com.example.termprojectuser.Entity
+
+import com.example.termprojectuser.Entity.Menu
+
+data class RewardMenu(
+        val menu: Menu,
+        val point: Int
+)
