+++ app/src/main/java/com/example/termprojectuser/FirebaseHelper/FirebaseQueueIDHelper.kt
@@ -1,4 +1,4 @@
-package com.example.termprojectuser
+package com.example.termprojectuser.FirebaseHelper
 
 import com.google.firebase.database.DataSnapshot
 import com.google.firebase.database.DatabaseError
