--- app/src/main/java/com/example/termprojectuser/Queue.kt
@@ -1,6 +1,0 @@
-package com.example.termprojectuser
-
-data class Queue(
-    val orderList: Map<String, OrderList> = mapOf(),
-    val queueId: String = "Unknown"
-)
