--- app/src/main/java/com/example/termprojectuser/RewardMenu.kt
@@ -1,20 +1,0 @@
-package com.example.termprojectuser
-
-data class RewardMenu(
-        val menu: Menu,
-        val point: Int
-){
-//    companion object{
-//        fun createMenu(): ArrayList<RewardMenu>{
-//            return arrayListOf(
-//                    RewardMenu(Menu(R.drawable.yuzu_refresher, "Yuzu refresher", 35,10), 40),
-//                    RewardMenu(Menu(R.drawable.classic_brown_sugar_milk_tea,"Classic brown sugar milk tea", 75,10 ), 150),
-//                    RewardMenu(Menu(R.drawable.matcha_brown_sugar_latte, "Matcha brown sugar latte", 105,10), 130),
-//                    RewardMenu(Menu(R.drawable.traditional_thai_milk_tea, "Traditional Thai milk tea", 15,10), 110),
-//                    RewardMenu(Menu(R.drawable.caramel_macchiato, "Caramel macchiato", 40,10), 100)
-//            )
-//
-//        }
-//    }
-
-}
